requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
albumentations==1.3.1
async-timeout==4.0.3
attrs==23.1.0
aug==0.2.4
brotlipy==0.7.0
certifi==2023.7.22
cffi 
charset-normalizer
click
cloudpickle
cmake==3.27.7
contourpy
cryptography
cycler
cytoolz
dask
datasets==2.14.4
dill==0.3.7
docopt==0.6.2
et-xmlfile==1.1.0
filelock==3.12.4
fonttools==4.25.0
frozenlist==1.4.0
fsspec
huggingface-hub==0.16.4
idna
imagecodecs
imageio
img==2.5
imgaug
importlib-metadata
importlib-resources
Jinja2==3.1.2
joblib==1.3.2
kiwisolver
Levenshtein==0.21.1
lightning-utilities==0.9.0
lit==17.0.2
lmdb==1.4.1
locket
lpips==0.1.4
MarkupSafe==2.1.3
matplotlib
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
munch==4.0.0
munkres==1.1.4
networkx
nltk==3.8.1
nougat-ocr==0.1.1
numpy==1.24.4
opencv-python==4.8.0.74
opencv-python-headless==4.8.0.76
openpyxl==3.1.2
orjson==3.9.5
packaging
pandas==1.4.2
partd
Pillow
pipreqs==0.4.13
platformdirs
pooch
pyarrow==13.0.0
pycparser
PyMuPDF==1.23.2
PyMuPDFb==1.23.0
pyOpenSSL 
pyparsing
PySocks
python-dateutil
python-Levenshtein==0.21.1
pytorch-lightning==2.0.7
pytz==2023.3
PyWavelets
PyYAML
qudida==0.0.4
rapidfuzz==3.2.0
regex==2023.8.8
requests
ruamel.yaml==0.17.32
ruamel.yaml.clib==0.2.7
scikit-image 
scikit-learn==1.3.0
scipy==1.10.1
sconf==0.2.5
sentencepiece==0.1.99
setproctitle==1.3.2
shapely
six 
sympy==1.12
thop==0.1.1.post2209072238
threadpoolctl==3.2.0
tifffile 
timm==0.5.4
tokenizers==0.13.3
toolz
torch==2.0.0+cu118
torchaudio==2.0.1+cu118
torchmetrics==1.1.1
torchvision==0.15.1+cu118
tqdm==4.66.1
transformers==4.25.1
triton==2.0.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3 
xlrd==1.2.0
xxhash==3.3.0
yarg==0.1.9
yarl==1.9.2
zipp
